requirements.txt
azure-storage-blob
fastapi
uvicorn
python-multipart
python-dotenv
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
rank-bm25
PyMuPDF
numpy
pydantic

# azure-storage-blob==12.27.1
# fastapi==0.122.0
# uvicorn==0.38.0
# python-multipart==0.0.20
# python-dotenv==1.2.1
# langchain-openai==1.1.0
# langchain-community==0.4.1
# langchain-core==1.1.0
# langchain-text-splitters==1.0.0
# faiss-cpu==1.13.0
# rank-bm25==0.2.2
# PyMuPDF==1.26.6
# numpy==2.2.6
# pydantic==2.12.5



# fastapi
# uvicorn
# python-dotenv
# azure-storage-blob
# langchain
# langchain-openai
# langchain-community
# pymupdf
# rank-bm25
# numpy
# tiktoken
# faiss-cpu
# clear
# azure-storage-blob
# uvicorn
# python-multipart
# pydantic
# uuid
